+++ src/common/Notes/const.js
@@ -1,0 +1,9 @@
+export const PANE_HEADER_APP_ICON = 'orders';
+
+export const ENTITY_TYPE_TRANSLATION_KEYS = {
+  poLine: 'ui-orders.poLine',
+};
+
+export const ENTITY_TYPE_PLURALIZED_TRANSLATION_KEYS = {
+  poLine: 'ui-orders.poLine.pluralized',
+};