+++ src/common/Notes/index.js
@@ -1,3 +1,1 @@
-export { default as NoteCreate } from './NoteCreate';
-export { default as NoteEdit } from './NoteEdit';
-export { default as NoteView } from './NoteView';
+export { default as Notes } from './Notes';